% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggweekly.R
\docType{data}
\name{us_federal_holidays}
\alias{us_federal_holidays}
\title{U.S. Federal Holidays}
\format{A data frame with 100 rows and 5 variables:
\describe{
\item{\code{year}}{character. Year}
\item{\code{day}}{double. Date}
\item{\code{label}}{character. Holiday label}
\item{\code{color}}{character. Color for text}
\item{\code{fill}}{character. Fill color}
}}
\usage{
us_federal_holidays
}
\description{
Federal Holidays from \url{https://opm.gov}
}
\references{
\url{https://www.opm.gov/policy-data-oversight/snow-dismissal-procedures/federal-holidays}
}
\keyword{datasets}
