% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggweekly.R
\name{ggweek_planner}
\alias{ggweek_planner}
\title{Create a Weekly Planner}
\usage{
ggweek_planner(start_day = lubridate::today(), end_day = start_day +
  lubridate::weeks(8) - lubridate::days(1), highlight_days = NULL,
  week_start_label = c("month day", "isoweek", "none"),
  show_day_numbers = TRUE, show_month_start_day = TRUE,
  show_month_boundaries = TRUE, highlight_text_size = 2,
  month_text_size = 4, month_color = "#f78154",
  holidays = ggweekly::us_federal_holidays,
  font_label_text = "PT Sans Narrow")
}
\arguments{
\item{start_day}{The starting day, either as a YYYY-mm-dd text string or
using \code{\link[lubridate:ymd]{lubridate::ymd()}}.}

\item{end_day}{The ending day, either as a YYYY-mm-dd text string or
using \code{\link[lubridate:ymd]{lubridate::ymd()}}.}

\item{highlight_days}{Special days to be highlighted, formatted as a tibble
with columns \code{day}, \code{label}, \code{color}, and \code{fill}.}

\item{week_start_label}{What labels should be used for the week start, i.e.
to the left of the first day of the week? One of \code{"month day"},
\code{"isoweek"}, or \code{"none"}.}

\item{show_day_numbers}{Should day numbers be included in each box of the
calendar?}

\item{show_month_start_day}{Should the first day of the month be highlighted?}

\item{show_month_boundaries}{Should boundaries between months be highlighted
with a dotted line?}

\item{highlight_text_size}{Size of the text in the highlighted days.}

\item{month_text_size}{The size of the text labelling the first day of the month.}

\item{month_color}{The color of the boxes highlighting the first day of the month.}

\item{holidays}{A tibble containing holiday dates in the same format as
\code{higlight_days}. Defaults to a list of US Federal Holidays. Set to \code{NULL}
to disable.}

\item{font_label_text}{Font for label text, default is
\href{https://fonts.google.com/specimen/PT+Sans+Narrow}{PT Sans Narrow}.}
}
\description{
Creates a weekly calendar with ggplot2
}
